chore: cleanup logging and make it configurable (https://github.com/snapshot-labs/sx-api/pull/30)
This cleans up logging within the checkpoint library and adds a configuration option to the Checkpoint class's initialization. This way users can easily turn on logs (and specify what level) if they are interested in monitoring the actions performed within the library.

This also moves the `graphql` handler from being initialized within Checkpoint's constructor into a getter function. This reduces the number of initialization being done when constructing a Checkpoint object and allows the user the freedom to initialize the graphql handler only when they need it.

In addition to moving the graphql handlers initialization, this also passes the logger to the resolver via context, so logging can be controller there as well.