refactor: simple quorum comment and var name (snapshot-labs/sx-evm#246)
The natspec comment mentioned that the quorum was calculated from all vote types when actually it is calculated from just for and abstain votes. 

Additionally, in the `_quorumReached` internal function, `totalVotes` was used as the variable name for the sum of for and abstain votes. This is misleading. 

This PR therefore updates the comment and renamed the `totalVotes` var to `forAndAbstainVotesTotal`.