refactor: replace choice magic numbers with string values (https://github.com/snapshot-labs/sx-ui/pull/861)
We used to use 1 | 2 | 3 for vote choice on UI (for, against, abstain).
Which was inconsistent with sx.js (0 abstain, 1, for, 2 against).

It could lead to issues as those values are not really clear on the UI.
This commit replaces 1 | 2 | 3 with 'for' | 'against' | 'abstain'
that are converted to SDK values in single place. This makes UI code
more readable as actions are easy to understand.