feat: handle special case of N+1
If DEFAULT_MAX_CHOICES is 6 and there are 7 choices
we would display "Others" button for just single choice, that would
later turn into real choice on click.

This commit changes it so normally if there are more choices we will
show 6 real choices + "Others" button, but if there are 7 choices
we will show 7 real choices and no "Others" button.