fix: use contract to map timestamp to block number
Previously we used Herodotus API to read block number for given timestamp.
This is not ideal because it doesn't actually mean that this mapping is already
available on the network (because mana needs time to process it and create transaction
out of it).

This commit removes dependency on Herodotus API on UI. Now mapping
timestamp to block number is done on strategy contract directly.
If it resolves it means that it will also be available for contract
when casting a vote or computing VP. If it's not available we are
marking it as NOT_READY_YET so UI displays appropriate message.