feat: add metadata network constant for aliases/follows/profiles
Currently aliases/follows/user profiles are fetched for first
offchainNetwork that is enabled.

This is a bit annoying if you want to display both networks at the same
time (in development or in sx-devnet deployment) because you need to switch
env and metadata network is dependent on the order in your ENABLED_NETWORKS.

With this change we separate enabled networks from the network that is
used for metadata and enable both s and s-tn networks at the same time.