ci: deploy to staging mainnet API app
With new script deployment to DO app will be decided based on the URL
that api.snapshot.box currently points to.

We use 3dns.box's authoritative nameserver to get latest records.

If it's currently pointing to APP_1 we will deploy to APP_2,
if it's currently pointing to APP_2 we will deploy to APP_1.