fix: prevent fetching proposal on wrong network
Currently we watch both networkId + spaceAddress and id.
But those values are not updated in-sync (id) is synchronous and
networkId + spaceAddress are resolved async.

This means that if URL changes from one proposal to another we will call:
1. fetchProposal(oldNetwork, oldSpaceAddress, newProposalId)
2. fetchProposal(newNetwork, newSpaceAddress, newProposalId)

First call is invalid and it shouldn't happen.

This commit changes our watcher to only call fetchProposal
when async properties change and id is read from current state.