fix: watch proposal id, but ignore if space is unresolved
Because space data and proposal ID are resolved in different stages
we have to avoid fetching new proposal if we have mismatched data
(proposalId from new space, space data from new space).

Previously I changed it so we only watch space data and read proposalId
on demand, but it introduced a bug that if only proposalId changes
it wouldn't be triggered at all.

Now we watch both space data and proposalId, but only act when we know
that space data was resolved.