fix: ceil scrollLeft
For some reason scrollBy might not scroll all the way (for example 199.5 instead of 200).
Further calls won't cause another scroll so we might be missing 0.5 to the right.

This changes it so if we are off by less than 1 pixel we still treat it as scrolled
all the way to the right.