feat: add explore page Category and Network filters (#882)
* feat: add signle category filter on explore page

* feat: add network selector on Explore page

* fix: render dropdown below header

* feat: use combobox for network selection on explore page

* fix: restore title on non-inline Comboboxes

* fix: use multi-line layout if needed

* fix: select contents on next frame

With DevTools it looks like when you select text right-click menu is triggered automatically.
This happens on all websites, we can prevent it on initial click by delaying it
to next frame, but if you click again it will be triggered anyway.

Not sure if this is worth keeping, it doesn't cause real-world issues AFAIK.

* feat: use gap similar to Dropdown to match

* feat: allow bigger width dropdown with truncate

* feat: resize input to match contents

* fix: do not trigger button click on space

* chore: update no results message

* refactor: use ghost text to resize input

* feat: sort by usage

* fix: only show testnet snapshot networks if metadataNetwork is s-tn