fix: separate display and cumulative decimals
In case of onchain strategies decimals are only there to specify how
strategy value should be formatted, but it doesn't have effect on how
this value is interpreted (strategy with VP value of 1 and 18 decimals and
strategy with VP value of 1 and 0 decimals are actually the same and give sum
of 2 VP).

We need to separate cumulativeDecimals (used to calculate cumulative VP) and
displayDecimals (used to display specific strategy's value).

Depending on the context one or another needs to be used.