fix: make proposals queries reactive by accepting MaybeRefOrGetter (#1190)
Previously our proposals queries were not reactive as they should, because
we only accepted rawValues. However we should make those values refs,
so queries are properly reactive.

We do this by following composables recommendation and making all values accept
MaybeRefOrGetter: https://vuejs.org/guide/reusability/composables.html#input-arguments

Lack of this reactivity caused bunch of issues as we migrated to use vue-query
for Spaces. This might not have any effect in current state (so it's more of a refactor
now), but is needed for resubmit of that PR.