feat: use multichain flow to show QR code by default
There are various edge cases that are hard to handle because
we can't await AppKit methods as it's event based. Trying to handle
them through events we have access to (state.initalized) isn't enough
as well. The most reliable way of handling it was actually adding sleep.
It's ugly, but it seems to work. I don't think we can do better
unless AppKit exposes more information.