fix: use unique ID for strategies metadata
For proposal validation startegies metadata we take last N items
as current.

If we use try to reuse existing entities it can result in issues, for example:
1. Add whitelist strategy (ID: 0x001/0/ipfs-hash-1).
2. Remove whitelist strategy and add ERC20 strategy (ID: 0x001/0/ipfs-hash-2).
3. Remove ERC20 strategy and add *same* whitelist strategy (ID: 0x001/0/ipfs-hash-1).

Because we already had that ID we won't create it, so
voting_power_validation_strategies_parsed_metadata would have 0x001/0/ipfs-hash-2
as last entry, which isn't correct.