fix: compute maxIndex from last entry in strategies metadatas
Previously we computed maxIndex just by looking at entire list of entries, but
this is not correct.

For example if in the past you had 3 strategies and now only have 1
maxIndex would end up being at 2 (because it's max index in all entries),
but it doesn't matter for current state.

We can compute it from last item which is now expected to be also
inserted last.