fix: make strategies order stable (#1314)
* fix: use unique ID for strategies metadata

For proposal validation startegies metadata we take last N items
as current.

If we use try to reuse existing entities it can result in issues, for example:
1. Add whitelist strategy (ID: 0x001/0/ipfs-hash-1).
2. Remove whitelist strategy and add ERC20 strategy (ID: 0x001/0/ipfs-hash-2).
3. Remove ERC20 strategy and add *same* whitelist strategy (ID: 0x001/0/ipfs-hash-1).

Because we already had that ID we won't create it, so
voting_power_validation_strategies_parsed_metadata would have 0x001/0/ipfs-hash-2
as last entry, which isn't correct.

* fix: compute maxIndex from last entry in strategies metadatas

Previously we computed maxIndex just by looking at entire list of entries, but
this is not correct.

For example if in the past you had 3 strategies and now only have 1
maxIndex would end up being at 2 (because it's max index in all entries),
but it doesn't matter for current state.

We can compute it from last item which is now expected to be also
inserted last.

* chore: bump @snapshot-labs/checkpoint to 0.1.0-beta.57