feat: display proposals without metadata (#1316)
* feat: display proposals without metadata

This PR changes how we handle proposals without metadata.

Now they will be dispalyed with placeholder values.

At the same time it will be marked as invalid and appropriate label
and message will be displayed on the UI.

If proposal also includes execution different messaging is used prompting
users to reject it.

* chore: update invalid proposal with execution message

* chore: use icon instead of FLAGGED label

* chore: use exclamation triangle icon