test: add typecheck to sx.js tests (#1320)
* test: add typecheck to sx.js tests

Currently we ignored test directory when building so even though
tests are passing we have lots of type errors in test directory.

To address this new tsconfig.json is added that is used in development
and for typecheck (this is also used in editor).

We have two variants of build TS configs for CJS and ESNext builds.

* chore: increase timeout for Starknet setup

* test: quit easrly on integration test failure

* test: get execution data in beforeAll