chore: run dev:interactive script with node (#1332)
Currently we run that script with ts-node, but it has its own
latency so it takes a while for our code to run.

Node v22.6.0 shipped with TS type stripping (behind experimental flag).
It's no longer exerimental in Node v23, but we keep those flags for now.

By using node directly we make this script startup faster.