chore: use esnext tsconfig for building dev version of sx.js
As our default config (used in IDEs) includes both src/ and test/
then building project with it will create those subdirectories in dist/

But this is not the structure our npm package exposes (files should be
in dist directly). Because of this UI won't automatically reload when
there are changes in sx.js.