refactor: support null current on EVM (#1392)
* refactor: support null current on EVM

For some reason we didn't support null current on EVM, but revisting it
I don't see reason why that would be the case.

This is needed however to implement apeGas voting strategy.

All strategies we use on EVM use block number just as blockTag so
null current can be replaced with "latest" blockTag.

This is likely very old limitation that we had initially that we kept
instead of normalizing.

To make sure it works I added tests that test null block number for all
strategies. Those tests are mocked now (because latest block can change
if balance changes), but as of now they should also work without mocks.

* refactor: simplify getProposalSnapshot