feat: make ApeGas proposals show up as pending for 2 hours. (#1406)
* refactor: only create apevote if snapshot already passed

This will handle case if voting delay is set. We can't create
apevote request with Herodotus if snapshot is in the future.

* feat: add 2 hour delay to ApeGas proposals

We need additional 2 hour delay for ApeGas proposals.
It will be added on top of existing delay as we need to wait from the
point where snapshot block is minted.

* fix: remove doubled infinite loop