fix: separate completed and execution_settled
Originally `completed` was used only for onchain spaces and was used for
execution, but later on it was reused for pending results
(https://github.com/snapshot-labs/sx-monorepo/pull/1055).

This meant that it started carrying two very different meanings and
broke things, so later on completed for onchain spaces was updated
to match offchain space meaning
(https://github.com/snapshot-labs/sx-monorepo/pull/1075) which broke
onchain execution.