chore: add execution_settled field
This will replace completed in the future. Renaming fields in API
are annoying because we need to carry two in the API during migration.
For now we are just adding new field, but still using old one in UI.

In the future next time we update API we can just remove completed field
and start using execution_settled on UI.