fix: separate completed and execution_settled (#1444)
* fix: separate completed and execution_settled

Originally `completed` was used only for onchain spaces and was used for
execution, but later on it was reused for pending results
(https://github.com/snapshot-labs/sx-monorepo/pull/1055).

This meant that it started carrying two very different meanings and
broke things, so later on completed for onchain spaces was updated
to match offchain space meaning
(https://github.com/snapshot-labs/sx-monorepo/pull/1075) which broke
onchain execution.

* chore: add execution_settled field

This will replace completed in the future. Renaming fields in API
are annoying because we need to carry two in the API during migration.
For now we are just adding new field, but still using old one in UI.

In the future next time we update API we can just remove completed field
and start using execution_settled on UI.