ci: share concurrency between promote and deploy
There is potential issue of conflict between deploy and promote
workflows.

For example, if there is new deployment workflow currently running
and we start promote workflow it's possible that we will promote
to app instance that is about to get reset.

Similar issue is possible if promote workflow starts first
and deploy runs next.

To avoid this we can put those two workflows into the same concurrency
group. With this in place following cases will be possible.

1. Promote workflow is running, deploy workflow wants to start - deploy
   workflow will only start once promtoe workflow is finished.
2. Deploy workflow is running, promote workflow wants to start - promote
   workflow will only start once deploy workflow is finished.