feat: deploy and promote in same concurrency context (#1509)
* ci: share concurrency between promote and deploy

There is potential issue of conflict between deploy and promote
workflows.

For example, if there is new deployment workflow currently running
and we start promote workflow it's possible that we will promote
to app instance that is about to get reset.

Similar issue is possible if promote workflow starts first
and deploy runs next.

To avoid this we can put those two workflows into the same concurrency
group. With this in place following cases will be possible.

1. Promote workflow is running, deploy workflow wants to start - deploy
   workflow will only start once promtoe workflow is finished.
2. Deploy workflow is running, promote workflow wants to start - promote
   workflow will only start once deploy workflow is finished.

* ci: check last deployed commit ID

Before we only looked at current commit, but it was unreliable.

For example if we had:
1. Commit that changes api (workflow starts)
2. Commit that doesn't change api (previous workflow gets cancelled, new
   one starts).
3. That new workflow wouldn't see changes.

With this change we compare actually deployed version.