feat: add Queued & Vetoed proposal status (#1515)
* feat: add Queued & Vetoed proposal status

* refactor: make useExecutionActions reactive

Previously useExecutionActions was not reactive.
If proposal changed (either to different ones or same one but updated)
it would be stuck with old data.

That meant that for example if you execute proposal and navigate away
and back to the UI proposal would be in Queued state, but execution
actions would be stuck in the previous state.

This addresses this issue.

* fix: remove typo in rejected