test: set up e2e tests using Playwright (#1497)
* test: set up e2e tests using Playwright

Right now it only tests UI (with Metamask), but in the future
it will set up more services (Anvil with deployed contracts, local API,
local Mana) to be more independent and reproducible.

* feat: add e2e tests to CI

* test: use playwright container to avoid downloading browsers

* chore: cleanup export

* test: add cache support for setup-node

* chore: clean up locators

* test: add space tests

* test: add ModalHelper POM

* test: add AuthPage POM

* test: normalize test names

* test: check if space can be opened from explore page

* chore: increase expect timeout to 10 seconds

* fix: only try to download extension in main worker

Right now each worker tries to download extension and they conflict with
each other.

With this change only main worker downloads extension and the rest wait
for it to be downloaded.

* ci: store videos on CI

* ci: separate playwright-report and videos artifacts

* ci: fix videos dir path

* ci: use traces instead of videos

videos are not recorded properly when reusing context