chore: do not save new apegas proposals
Herodotus API is down so all of those will fail.

We also have issue with storing duplicates of those in DB which
needs to be fixed by adding proper unique index, but we need to
clean up DB before we can add which would be hard if API can write
more entries.