test: use raw dispatch event on space item (#1671)
* test: use raw dispatch event on space item

We started using new layout which shows Follow button on hover.
With the way Playwright works it interprets it as element that would
actually be clicked by the user.

We want to click on the actual space button (and assume that Follow
button won't actually obstruct the click).

Caused by change in: https://github.com/snapshot-labs/sx-monorepo/pull/1626

* test: increase starknet integration tests timeout to 120s

Looks like we are right at the limit of 60s right now so it's flaky.