refactor: interpret VITE_WHITELABEL_MAPPING as forced whitelabel (#1701)
* refactor: interpret VITE_WHITE_LABEL_MAPPING as forced whitelabel

Before even if VITE_WHITE_LABEL_MAPPING was defined we only enabled
whitelabel if we visited UI from non-default domain (so localhost
would be non-whitelabel and 127.0.0.1 was white label).

This is not super clear, because even if we run it on 127.0.0.1
then VITE_HOST is still localhost.

This also meant that vite dev script was returning 127.0.0.1 URL
which we rarely use.

To avoid that now when VITE_WHITE_LABEL_MAPPING is used we just enable
whitelabel straight away. This way script runs on localhost directly
and code should be a bit easier to understand as well.

* chore: rename WHITE_LABEL to WHITELABEL