refactor: schedule nextTick explicitly instead of magic await
In order and payment flows we currently use await to schedule
re-opening of the modal after next tick. This is however done implicitly
by just adding await to sync function which triggers linter (and other
developers' eyes) so it's confusing and at risk of being removed in the
future.

Instead we just explicitly use nextTick to schedule modal to re-open
after next render when currentStep is already updated.