refactor(auction): store sell orders as sellAmount/buyAmount
To reduce dependence on the price (which is not used on the contracts)
we store sell orders as { sellAmount, buyAmount } where both of them
are bigints representing raw units of applicable tokens.

We convert to this format as soon as possible and use those amounts in
entire flow (with exception of fetching previous order from API as there
is no other way, however this price is computed based on those values
instead of being user input). This will be further improved in https://github.com/snapshot-labs/sx-monorepo/pull/1799