fix: fix sell price precision (#1799)
* fix: fix sell price precision

* refactor: simplify import path

* fix: fix yarn.lock changes

* refactor(auction): store sell orders as sellAmount/buyAmount (#1800)

To reduce dependence on the price (which is not used on the contracts)
we store sell orders as { sellAmount, buyAmount } where both of them
are bigints representing raw units of applicable tokens.

We convert to this format as soon as possible and use those amounts in
entire flow (with exception of fetching previous order from API as there
is no other way, however this price is computed based on those values
instead of being user input). This will be further improved in https://github.com/snapshot-labs/sx-monorepo/pull/1799

* fix(auction): align order sorting with contracts (#1801)

* fix: update order comparison criteria

* refactor: reverse price comparison order to be easier to understand

Defining price as sellAmount/buyAmount.

* fix: fix sell price precision

* fix: fix sell price precision

* refactor: simplify import path

* refactor: better variable name

* fix: fix yarn.lock

---------

Co-authored-by: Wiktor Tkaczyński <wiktor.tkaczynski@gmail.com>