feat: add helper to compute buyAmount with token precision (#1803)
* feat: add helper to compute buyAmount with token precision

This PR adds helper (getOrderBuyAmount) that computes accurate
buyAmount for order considering the given sellAmount and price.

This helper retains full precision of the token and ensures values are
rounded up instead of truncated (for example price of 0.3 results in
buyAmount of 3.333334 instead of 3.333333 (which makes actual price
lower than expected instead of higher).

* refactor: use parseUnits in toRawUnits